+++ test/controllers/application_controller_test.rb
@@ -2,45 +2,54 @@
 
 require_relative '../test_helper'
 require 'minitest/autorun'
-
-class ApplicationController
-  attr_reader :current_user
-
-  def initialize(user)
-    @current_user = user
-  end
-
-  def head(status) = status
-
-  def update_location(params)
-    @current_user.update!(lat: params[:lat], long: params[:long]) if @current_user && params[:lat] && params[:long]
-    head :ok
-  end
-
-  def know_tree(params)
-    if @current_user && (tree = params[:tree])
-      @current_user.known_trees ||= []
-      @current_user.known_trees << tree unless @current_user.known_trees.include?(tree)
-    end
-    head :ok
-  end
-end
+require_relative '../../app/controllers/application_controller'
 
 class ApplicationControllerTest < Minitest::Test
   def test_update_location_updates_user
     user = User.new
-    controller = ApplicationController.new(user)
-    controller.update_location(lat: 1.5, long: 2.5)
+    controller = ApplicationController.new
+    controller.instance_variable_set(:@current_user, user)
+    controller.params = { lat: 1.5, long: 2.5 }
+    controller.update_location
     assert_equal 1.5, user.lat
     assert_equal 2.5, user.long
   end
 
   def test_know_tree_adds_tree_to_user
-    user = User.new
-    user.define_singleton_method(:known_trees) { @known_trees ||= [] }
-    tree = Tree.new
-    controller = ApplicationController.new(user)
-    controller.know_tree(tree: tree)
-    assert_includes user.known_trees, tree
+    user = User.new(id: 1)
+    tree = Tree.new(id: 2)
+
+    Tree.singleton_class.class_eval do
+      attr_accessor :records
+      def find_by(id:)
+        Array(records).find { |t| t.id == id }
+      end
+    end
+
+    UserTree.singleton_class.class_eval do
+      attr_accessor :records
+      def find_or_create_by!(user:, tree:)
+        self.records ||= []
+        rec = records.find { |r| r[:user] == user && r[:tree] == tree }
+        unless rec
+          rec = { user: user, tree: tree }
+          records << rec
+        end
+        rec
+      end
+    end
+
+    Tree.records = [tree]
+    UserTree.records = []
+
+    controller = ApplicationController.new
+    controller.instance_variable_set(:@current_user, user)
+    controller.params = { id: 2 }
+    controller.know_tree
+
+    assert_includes UserTree.records, { user: user, tree: tree }
+  ensure
+    Tree.records = nil
+    UserTree.records = nil
   end
 end